🚚 Moved old pages structure temp to components
