+++ next.config.js
@@ -1,5 +1,2 @@
 module.exports = {
-    experimental: {
-        appDir: true,
-    },
 }
