+++ tsconfig.json
@@ -21,9 +21,9 @@
         "baseUrl": ".",
         "paths": {
             "@/components/*": [
-                "components/*",
+                "oldPages/_home/*,
             ],
-            "@/imgs/*": [
+   oldPages/_home/aboutMe/*[
                 "public/imgs/*"
             ],
             "@/svgs/*": [
